requirements.txt
tensorflow==1.14.0
Pillow==7.1.2
Keras==2.3.1
numpy